requirements.txt
streamlit==1.32.0        # For the frontend interface
pandas==2.2.1            # For data handling and processing
google-generativeai==0.4.1  # For Gemini API integration (version may vary, check latest)
python-dotenv==1.0.1     # For loading environment variables (e.g., API keys)